updates db integration tests to work with weaviate
